+++ hydra-cluster/test/Test/OfflineChainSpec.hs
@@ -11,7 +11,7 @@ import Data.Aeson qualified as Aeson
 import Data.Aeson.Lens (key, _Number)
 import Hydra.Cardano.Api (ChainPoint (..), Tx, UTxO)
 import Hydra.Chain (ChainCallback, ChainEvent (..), ChainStateHistory, OnChainTx (..), initHistory)
-import Hydra.Chain.Direct.State (initialChainState)
+import Hydra.Chain.Direct.State (ChainStateAt (..), initialChainState)
 import Hydra.Chain.Offline (withOfflineChain)
 import Hydra.Cluster.Fixture (alice)
 import Hydra.Cluster.Util (readConfigFile)
@@ -55,7 +55,7 @@ spec = do
       withOfflineChain offlineConfig alice [] noHistory callback $ \_chain -> do
         -- Expect to see a tick of slot 1 within 2 seconds
         waitMatch waitNext 2 $ \case
-          Tick{chainPoint = ChainPoint chainSlot _} -> guard $ chainSlot > 0
+          Tick{chainState = ChainStateAt{recordedAt = Just (ChainPoint chainSlot _)}} -> guard $ chainSlot > 0
           _ -> Nothing
 
   it "does not start on slot 0 with real genesis file" $ do
@@ -73,7 +73,7 @@ spec = do
       withOfflineChain offlineConfig alice [] noHistory callback $ \_chain -> do
         -- Should not start at 0
         waitMatch waitNext 1 $ \case
-          Tick{chainPoint = ChainPoint chainSlot _} -> guard $ chainSlot > 1000
+          Tick{chainState = ChainStateAt{recordedAt = Just (ChainPoint chainSlot _)}} -> guard $ chainSlot > 1000
           _ -> Nothing
         -- Should produce ticks on each slot, which is defined by genesis.json
         Just slotLength <- readFileBS (tmpDir </> "genesis.json") >>= \bs -> pure $ bs ^? key "slotLength" . _Number