+++ hydra-node/src/Hydra/Chain/Direct.hs
@@ -44,10 +44,11 @@ import Hydra.Chain (
   ChainComponent,
   ChainStateHistory,
   PostTxError (..),
-  latestKnownChainPoint,
+  lastKnown,
  )
 import Hydra.Chain.Backend (ChainBackend (..))
 import Hydra.Chain.CardanoClient qualified as CardanoClient
+import Hydra.Chain.ChainState (chainStatePoint)
 import Hydra.Chain.Direct.Handlers (
   CardanoChainLog (..),
   ChainSyncHandler,
@@ -138,7 +139,7 @@ withDirectChain ::
   ChainComponent Tx IO a
 withDirectChain backend tracer config ctx wallet chainStateHistory callback action = do
   -- Last known point on chain as loaded from persistence.
-  let persistedPoint = Just (latestKnownChainPoint chainStateHistory)
+  let persistedPoint = Just (chainStatePoint $ lastKnown chainStateHistory)
   queue <- newLabelledTQueueIO "direct-chain-queue"
   -- Select a chain point from which to start synchronizing
   chainPoint <- maybe (queryTip backend) pure $ do